+++ packages/anon-aadhaar-contracts/interfaces/IAnonAadhaarVerifier.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.7.0 <0.9.0;
+
+interface IAnonAadhaarVerifier {
+    function verifyProof(
+        uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[34] calldata _pubSignals
+    ) external view returns (bool);
+}